Decoupled the IoMediator classes as much as possible.
- Cleaned up the imports. Most of the code interfacing with these classes now uses absolute imports
- Fixes #372
